Ser/requirements.txt
flask
flask-cors
pdfminer.six
docx2txt
gunicorn
google-generativeai
python-dotenv
